# YD-374 Optimize Gradle