--- a/batchservice/build.gradle
+++ b/batchservice/build.gradle
@@ -6,7 +6,7 @@ buildscript {
 	}
 	dependencies {
 		classpath("org.springframework.boot:spring-boot-gradle-plugin:$project.ext.springBootVersion")
-		classpath("io.spring.gradle:dependency-management-plugin:0.5.3.RELEASE")
+		classpath("io.spring.gradle:dependency-management-plugin:0.6.1.RELEASE")
 		classpath('se.transmode.gradle:gradle-docker:1.2')
 		classpath 'net.researchgate:gradle-release:2.3.4'
 	}
@@ -113,4 +113,8 @@ task intTest(type:Test){
 	description = "Run integration tests (located in src/intTest/...)."
 	testClassesDir = project.sourceSets.intTest.output.classesDir
 	classpath = project.sourceSets.intTest.runtimeClasspath
+}
+
+bootRepackage {
+        classifier = "full"
 }