--- a/build.gradle
+++ b/build.gradle
@@ -60,5 +60,5 @@ eclipse {
 }
 
 task wrapper(type: Wrapper) {
-	gradleVersion = '2.13'
+	gradleVersion = '3.1'
 }