--- a/dbinit/build.gradle
+++ b/dbinit/build.gradle
@@ -6,7 +6,7 @@ buildscript {
 	}
 	dependencies {
 		classpath("org.springframework.boot:spring-boot-gradle-plugin:$project.ext.springBootVersion")
-		classpath("io.spring.gradle:dependency-management-plugin:0.5.3.RELEASE")
+		classpath("io.spring.gradle:dependency-management-plugin:0.6.1.RELEASE")
 		classpath('se.transmode.gradle:gradle-docker:1.2')
 		classpath 'net.researchgate:gradle-release:2.3.4'
 	}
@@ -67,4 +67,8 @@ bootRun {
 		'spring.batch.initializer.enabled': "true",
 		'spring.jpa.hibernate.ddl-auto': "update"
 	]
+}
+
+bootRepackage {
+        classifier = "full"
 }