* Add qualifier to ensure everything is up-to-date after the build
* Upgraded to Gradle 3.1
* Upgraded io.spring.gradle:dependency-management-plugin from 0.5.3 to 0.6.1 (older version didn't work with Gradle 3.1)

FYI this is the performance difference:
Before applying this change set:
* Full build, after clean: 58 seconds
* Second build, right after the previous one: 34 seconds

After applying this change set:
* Full build, after clean: 58 seconds
* Second build, right after the previous one: 4.3 seconds

The major difference in incremental builds is caused by two things:
* The new Gradle version always uses the daemon and the new Gradle version is faster
* The build is truly incremental. Earlier, it used to rebuild the JARs and everything after that, due to a missing setting in the ``build.gradle`` files.